Integrate form and simulation modules
