+++ form/button_module.py
@@ -1,20 +1,26 @@
 import tkinter as tk
 from tkinter import ttk
+import threading
 
-def create_button_frame(parent_frame, name_entries, option, percentage_var):
+form_built_flag = threading.Event()
+
+def create_button_frame(parent_frame, name_entries, option, percentage_var, callback):
     frame = ttk.Frame(parent_frame, padding="10")
 
     def submit_form():
+        form_data = {}
         for name_label, name_entry in name_entries:
             name = name_entry.get()
-            print(f"{name_label}", name)
+            form_data[name_label[:-1]] = name
 
-        selected_option = option.get()
-        selected_percentage = percentage_var.get()
-        print("Selected Option:", selected_option)
-        print("Selected Percentage:", selected_percentage)
+        form_data["Selected Option"] = option.get()
+        form_data["Selected Percentage"] = percentage_var.get()
 
+        form_built_flag.set()
         parent_frame.destroy()
+
+        callback(form_data)
+        return form_data
 
     submit_button = ttk.Button(frame, text="OK", command=submit_form)
     submit_button.grid(row=0, column=0, columnspan=2, pady=5)
