+++ peripheral/constants.py
@@ -2,6 +2,8 @@ WIDTH, HEIGHT = 800, 600
 CAPTION = "Shoot Simulator"
 WHITE = (255, 255, 255)
 RED = (255, 0, 0)
+GREEN = (0, 255, 0)
+BLUE = (0, 0, 255)
 READING_ERROR = "Error Reading joycon:"
 INITIALIZATION_ERROR = "Error initializing external peripheral:"
 PRESS_STATUS = 1
